Add admin area. And make it ready for work.

+++ Calciolandia-Website/Controllers/ManagerController.cs
@@ -1,5 +1,7 @@
-﻿using Calciolandia_Website.Core.Constants;
+﻿using Calciolandia_Website.Areas.Admin;
+using Calciolandia_Website.Core.Constants;
 using Calciolandia_Website.Core.Contracts;
+using Calciolandia_Website.Core.Data.Models;
 using Calciolandia_Website.Core.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
@@ -27,6 +29,11 @@ namespace Calciolandia_Website.Controllers
         [Authorize(Roles = RoleConstants.Admin)]
         public async Task<IActionResult> Add()
         {
+            if (this.User.IsInRole(AdminConstants.AdminRoleName))
+            {
+                return RedirectToAction(nameof(Add), nameof(Manager), new { area = "Admin" });
+            }
+
             var model = new ManagerViewModel()
             {
                 FootballClubs = await managerService.GetFootballClubsAsync()
@@ -39,6 +46,11 @@ namespace Calciolandia_Website.Controllers
         [Authorize(Roles = RoleConstants.Admin)]
         public async Task<IActionResult> Add(ManagerViewModel model)
         {
+            if (this.User.IsInRole(AdminConstants.AdminRoleName))
+            {
+                return RedirectToAction(nameof(Add), nameof(Manager), new { area = "Admin" });
+            }
+
             if (!ModelState.IsValid)
             {
                 return View(model);
@@ -60,7 +72,12 @@ namespace Calciolandia_Website.Controllers
         [Authorize(Roles = RoleConstants.Admin)]
         public async Task<IActionResult> Delete(Guid id)
         {
-           await managerService.DeleteAsync(id);
+            if (this.User.IsInRole(AdminConstants.AdminRoleName))
+            {
+                return RedirectToAction(nameof(Delete), nameof(Manager), new { area = "Admin" });
+            }
+
+            await managerService.DeleteAsync(id);
 
             return RedirectToAction("Index", "Home");
         }
@@ -69,6 +86,10 @@ namespace Calciolandia_Website.Controllers
         [Authorize(Roles = RoleConstants.Admin)]
         public async Task<IActionResult> Edit(Guid id)
         {
+            if (this.User.IsInRole(AdminConstants.AdminRoleName))
+            {
+                return RedirectToAction(nameof(Edit), nameof(Manager), new { area = "Admin" });
+            }
             var model = await managerService.GetForEditAsync(id);
 
             model.FootballClubs = await managerService.GetFootballClubsAsync();
@@ -80,6 +101,10 @@ namespace Calciolandia_Website.Controllers
         [Authorize(Roles = RoleConstants.Admin)]
         public async Task<IActionResult> Edit(ManagerViewModel model)
         {
+            if (this.User.IsInRole(AdminConstants.AdminRoleName))
+            {
+                return RedirectToAction(nameof(Edit), nameof(Manager), new { area = "Admin" });
+            }
             if (!ModelState.IsValid)
             {
                 return View(model);
@@ -100,6 +125,10 @@ namespace Calciolandia_Website.Controllers
         [HttpGet]
         public async Task<IActionResult> Info(Guid id)
         {
+            if (this.User.IsInRole(AdminConstants.AdminRoleName))
+            {
+                return RedirectToAction(nameof(Info), nameof(Manager), new { area = "Admin" });
+            }
             var model = await managerService.GetManagerById(id);
 
             return View(model);
